requirements.txt
flask 
pyttsx3
SpeechRecognition
pandas
scikit-learn
nltk
joblib
matplotlib
transformers
datasets
evaluate
imblearn
requests
python-dotenv
sqlalchemy 
pymysql
pyaudio